+++ project-base/src/Shopsys/ShopBundle/Model/AdvancedSearch/Filter/ProductNameFilter.php
@@ -53,7 +53,7 @@ class ProductNameFilter implements AdvancedSearchFilterInterface
             if ($ruleData->value === null) {
                 $searchValue = '%';
             } else {
-                $searchValue = '%' . DatabaseSearching::getLikeSearchString($ruleData->value) . '%';
+                $searchValue = DatabaseSearching::getFullTextLikeSearchString($ruleData->value);
             }
             $dqlOperator = $this->getDqlOperator($ruleData->operator);
             $parameterName = 'productName_' . $index;